size

+++ src/fr/esisar/compilation/gencode/Generation.java
@@ -73,8 +73,8 @@ public class Generation {
         return index;
     }
 
-    private static int sizeTableau(Type type) {
-        int size = 1;
+    private static long sizeTableau(Type type) {
+        long size = 1;
         while (type.getNature() == NatureType.Array) {
             size = (type.getIndice().getBorneSup() - type.getIndice().getBorneInf() + 1) * size;
             type = type.getElement();
@@ -170,7 +170,7 @@ public class Generation {
         Registre result = memoire.get();
         Registre value = memoire.get(result);
         Registre temp = memoire.get(result, value);
-        int size = sizeTableau(emplacement_Variable(a.getFils1(), result));
+        int size = (int)sizeTableau(emplacement_Variable(a.getFils1(), result));
         emplacement_Variable(a.getFils2(), value);
         for (int i = 0; i < size; i++) {
             Prog.ajouter(Inst.creation2(Operation.LOAD, Operande.creationOpIndirect(i, value), Operande.opDirect(temp)));
@@ -194,7 +194,7 @@ public class Generation {
                 Type type = emplacement_Variable(a.getFils1(), registre);
                 Etiq erreur = Etiq.nouvelle("eti" + numberEti++);
                 Etiq fin = Etiq.nouvelle("eti" + numberEti++);
-                int size = sizeTableau(type.getElement());
+                int size = (int)sizeTableau(type.getElement());
                 Registre index = memoire.get(registre);
                 Operande operande = gene_Exp(a.getFils2(), index);
                 if (operande.getNature() != NatureOperande.OpDirect)
